requirements.txt
ipykernel
langchain_core
python-dotenv
langchain-community 
pypdf
goog
setuptools
faiss-cpu
langchain_objectbox
langchain_huggingface
sentence-transformers
PyPDF2
chromadb
langchain_text_splitters
langchain
openpyxl
langchain_openai
pyprojroot
langchain_chroma